More fixes to work around groovy and its rediculousness.
git-svn-id: svn+ssh://cvs1/svn/listen/trunk@3217 4692729f-32a0-46b7-9832-1bf53c4881b5

+++ controller/listen/grails-app/domain/com/interact/listen/attendant/action/Action.groovy
@@ -15,5 +15,5 @@ abstract class Action
         keysPressed nullable: true, blank: false, matches: '^[0-9\\?\\*#]+$' // question mark is a single-character wildcard
     }
 
-    abstract toIvrCommand(String promptDirectory, String promptBefore, String artifactsDirectory, int organizationId);
+    abstract toIvrCommand(String promptDirectory, String promptBefore, String artifactsDirectory, long organizationId);
 }
