+++ controller/listen/grails-app/domain/com/interact/listen/attendant/action/DialPressedNumberAction.groovy
@@ -3,7 +3,7 @@ package com.interact.listen.attendant.action
 class DialPressedNumberAction extends Action
 {
     // has no attributes
-    def toIvrCommand(String promptDirectory, String promptBefore, String artifactsDirectory, int organizationId)
+    def toIvrCommand(String promptDirectory, String promptBefore, String artifactsDirectory, long organizationId)
     {
         return [
             promptBefore: !promptBefore || promptBefore.trim() == '' ? '' : promptDirectory + '/' + promptBefore,
